Added flipping rendering and new gfx
GFraMe_sprite was modified to allow rendering a sprites flipped, while keeping the sprite's position relative to the hitbox the same. In order to do the flipping, it was necessary to modify GFraMe_texture, where the rendering actually happens and, thus, GFraMe spriteset and GFraMe_tilemap were modified accordingly.
I also redid the graphics using the DB 32 palette, and making the basic size 32x32