Added score tracking and displaying
A tilemap based score was added, as well as its functionallity to roll when increased (similar to Earthbound's hp). To make this (updating the score) easier, the buffer passed to tilemap is now stored as a reference, instead of copying it.
Also, the font itself was added to the sprite atlas.